Move conversion functions to definitions
